refactor waitForIdleNetwork as a helper method and import into POM
