+++ air-quality-ui/system_tests/pages/vAirify_city_page.ts
@@ -1,5 +1,7 @@
 import { type Locator, type Page } from '@playwright/test'
 
+import { waitForIdleNetwork } from '../utils/helper_methods'
+
 export class VairifyCityPage {
   readonly page: Page
   readonly title: Locator
@@ -22,7 +24,7 @@ export class VairifyCityPage {
   }
 
   async captureChartScreenshot() {
-    await this.waitForIdleNetwork()
+    await waitForIdleNetwork(this.page, this.aqiChart)
     return await this.aqiChart.screenshot()
   }
 
@@ -52,45 +54,5 @@ export class VairifyCityPage {
     await this.setupRioDeJaneiroRoute()
     await this.gotoRioCityPage()
     await this.waitForGraphVisible()
-  }
-
-  async waitForIdleNetwork() {
-    try {
-      await this.aqiChart.waitFor()
-
-      const idleTime = 1000
-      const checkInterval = 200
-      const timeout = 5000
-
-      let lastActivityTime = Date.now()
-
-      const checkNetworkIdle = async () => {
-        while (Date.now() - lastActivityTime < idleTime) {
-          await new Promise((resolve) => setTimeout(resolve, checkInterval))
-          const networkRequests = await this.page.evaluate(() => {
-            return performance
-              .getEntriesByType('resource')
-              .filter((entry: PerformanceEntry) => {
-                const resourceEntry = entry as PerformanceResourceTiming
-                return (
-                  resourceEntry.initiatorType === 'xmlhttprequest' ||
-                  resourceEntry.initiatorType === 'fetch'
-                )
-              }).length
-          })
-          if (networkRequests === 0) {
-            lastActivityTime = Date.now()
-          }
-        }
-      }
-
-      const timeoutPromise = new Promise((_, reject) =>
-        setTimeout(() => reject(new Error('Network idle timeout')), timeout),
-      )
-      await Promise.race([checkNetworkIdle(), timeoutPromise])
-    } catch (error) {
-      console.error('Error waiting for chart animation:', error)
-      throw error
-    }
   }
 }
